Basic interface to queue jobs. Can't see or interact with the queue beyond adding jobs, but it's a start. Added functions to handle sending various commands, but not currently using them. Updated .gitignore.

+++ IotDeviceManager/Models/SprinklerJob.cs
@@ -3,12 +3,12 @@ using System;
 public class SprinklerJob
 {
     public SprinklerJob() {}
-    public SprinklerJob(int zoneNumber, UInt64 duration_s)
+    public SprinklerJob(UInt16 zoneNumber, UInt64 duration_s)
     {
         ZoneNumber = zoneNumber;
         Duration_s = duration_s;
     }
 
-    public int ZoneNumber { get; set; }
+    public UInt16 ZoneNumber { get; set; }
     public UInt64 Duration_s {get; set; }
 }
